Created provider folder for passportSetup -Modified passportSetup to break it up into separate parts. -Modified app.js to require index.js instead of passportSetup, since it will also contain the references to passport, passportSerialize and to passportDeserialize. All three are necessary since passportSetup was broken up.
