+++ src/providers/passport/index.js
@@ -1,0 +1,3 @@
+require('./passportSetup');
+require('./passportSerialize');
+require('./passportDeserialize');
