+++ src/providers/passport/passportSerialize.js
@@ -1,0 +1,5 @@
+const passport = require('passport');
+
+passport.serializeUser((user, done) => {
+  done(null, user.id);
+});
